Add router-driven home page
